feat: Connect Maid Profile screen to live Firestore data
This commit connects the `MaidProfileScreen` to a live Firestore backend, replacing the placeholder data with real-time maid details and reviews. It also handles loading and error states and refines the navigation flow from the `MaidListScreen`.

### Core Changes:

-   **`MaidRepository.kt`**:
    -   A new `getMaidReviews` function is added to fetch reviews for a specific maid from the `/maids/{maidId}/reviews` subcollection in Firestore.

-   **`MaidProfileViewModel.kt`**:
    -   Refactored to use `MaidRepository` for fetching data.
    -   A new `loadMaid(maidId)` function orchestrates fetching both the maid's profile details and their reviews concurrently.
    -   It now manages UI states for loading, error, and success, updating the screen based on the results from Firestore.
    -   The `MaidProfile`, `CustomerReview`, and other data models have been updated to align with the Firestore data structure, replacing enums and placeholder types with primitive types like `Double` and `String`.

-   **`MaidProfileScreen.kt`**:
    -   Now accepts a `maidId` to trigger data loading via a `LaunchedEffect`.
    -   The UI dynamically displays a `CircularProgressIndicator` during data loading or an error message if the fetch fails.
    -   The content is rendered only after the `maidProfile` data is successfully loaded into the state.

### UI and Navigation:

-   **`MaidyNavHost.kt`**:
    -   Updated to correctly pass the `maidId` from the route to the `MaidProfileScreen`.

-   **`MaidListScreen.kt`**:
    -   The "Select" and "View Details" buttons now navigate to the `MaidProfileScreen`, passing the selected maid's ID.

-   **Component Updates**:
    -   **`BottomBookingBar.kt`**: Now handles an `isAvailable` state, disabling the "Book This Maid" button and changing its text to "Not Available" if the maid cannot be booked.
    -   **`ServicesOfferedSection.kt`** and **`ServiceItem.kt`**: Simplified to work with a `List<String>` of service names, removing the previous `ServiceIconType` enum.
    -   **`ReviewItem.kt`**: The reviewer's profile picture has been removed to simplify the component.

-   **Dependency Injection**:
    -   **`AppModule.kt`**: `MaidProfileViewModel` is now registered in the Koin dependency graph.

+++ app/src/main/java/com/example/maidy/feature/maid_details/components/CustomerReviewsSection.kt
@@ -36,7 +36,7 @@ fun CustomerReviewsSection(
                 reviewerName = review.reviewerName,
                 reviewerImageUrl = review.reviewerImageUrl,
                 date = review.date,
-                rating = review.rating,
+                rating = review.rating.toFloat(),
                 comment = review.comment
             )
 
@@ -57,7 +57,7 @@ fun CustomerReviewsSectionPreview() {
                         reviewerName = "Mark Johnson",
                         reviewerImageUrl = "",
                         date = "June 15, 2024",
-                        rating = 5f,
+                        rating = 5,
                         comment = "Elena was fantastic! Our house has never looked this clean. She was professional, punctual, and incredibly thorough. Highly recommend!"
                     ),
                     CustomerReview(
@@ -65,7 +65,7 @@ fun CustomerReviewsSectionPreview() {
                         reviewerName = "Sarah Lee",
                         reviewerImageUrl = "",
                         date = "June 12, 2024",
-                        rating = 5f,
+                        rating = 5,
                         comment = "Absolutely amazing service. Elena paid attention to every little detail. I'm so happy with the result and will definitely be booking her again."
                     )
                 )
@@ -73,5 +73,6 @@ fun CustomerReviewsSectionPreview() {
         }
     }
 }
+
